third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/link-onclick-entities-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Ca%20href=%27about:blank%27%20onclick=%27alert(String.fromCharCode(0x58,0x53,0x53))//%26amp%3Bcopy%3B%27%3EClick%3C/a%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
